# app/Models/CardSwitchTask.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
use Illuminate\Database\Eloquent\Relations\BelongsTo;
use Illuminate\Database\Eloquent\Factories\HasFactory;
/**
 * @OA\Schema(
 *      schema="CardSwitchTask",
 *      required={"uuid","card_id","merchant_id","status_id","user_id"},
 *      @OA\Property(
 *          property="uuid",
 *          description="",
 *          readOnly=false,
 *          nullable=false,
 *          type="string",
 *      ),
 *      @OA\Property(
 *          property="card_uuid",
 *          description="",
 *          readOnly=false,
 *          nullable=true,
 *          type="string",
 *      ),
 *      @OA\Property(
 *          property="previous_card_uuid",
 *          description="",
 *          readOnly=false,
 *          nullable=true,
 *          type="string",
 *      ),
 *      @OA\Property(
 *          property="merchant_uuid",
 *          description="",
 *          readOnly=false,
 *          nullable=true,
 *          type="string",
 *      ),
 *      @OA\Property(
 *          property="status_uuid",
 *          description="",
 *          readOnly=false,
 *          nullable=true,
 *          type="string",
 *      ),
 *      @OA\Property(
 *          property="user_uuid",
 *          description="",
 *          readOnly=false,
 *          nullable=true,
 *          type="string",
 *      ),
 *      @OA\Property(
 *          property="created_at",
 *          description="",
 *          readOnly=true,
 *          nullable=true,
 *          type="string",
 *          format="date-time"
 *      ),
 *      @OA\Property(
 *          property="updated_at",
 *          description="",
 *          readOnly=true,
 *          nullable=true,
 *          type="string",
 *          format="date-time"
 *      ),
 *      @OA\Property(
 *          property="deleted_at",
 *          description="",
 *          readOnly=true,
 *          nullable=true,
 *          type="string",
 *          format="date-time"
 *      )
 * )
 */
class CardSwitchTask extends Model
{
    use HasFactory;
    use SoftDeletes;

    public $table = 'card_switch_tasks';

    public $fillable = [
        'uuid',
        'card_uuid',
        'previous_card_uuid',
        'merchant_uuid',
        'status_uuid',
        'user_uuid',
        'card_id',
        'previous_card_id',
        'merchant_id',
        'status_id',
        'user_id'
    ];

    protected $casts = [
        'uuid' => 'string',
        'card_uuid' => 'string',
        'previous_card_uuid' => 'string',
        'merchant_uuid' => 'string',
        'status_uuid' => 'string',
        'user_uuid' => 'string'
    ];

    public static array $rules = [
        'card_id' => 'required',
        'merchant_id' => 'required',
    ];

    public function card(): BelongsTo
    {
        return $this->belongsTo(Card::class, 'card_id');
    }

    public function merchant(): BelongsTo
    {
        return $this->belongsTo(Merchant::class, 'merchant_id');
    }

    public function previousCard(): BelongsTo
    {
        return $this->belongsTo(Card::class, 'previous_card_id');
    }

    public function status(): BelongsTo
    {
        return $this->belongsTo(Status::class, 'status_id');
    }

    public function user(): BelongsTo
    {
        return $this->belongsTo(User::class, 'user_id');
    }
}